% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-ANIMALS.R
\name{ANIMALS}
\alias{ANIMALS}
\title{ANIMALS Test Scores}
\usage{
ANIMALS(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{ANIMALS}.
}
\description{
Create a \code{ANIMALS} object to hold ANIMALS scores.
}
