% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-BOSTON.R
\name{BOSTON}
\alias{BOSTON}
\title{BOSTON Test Scores}
\usage{
BOSTON(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{BOSTON}.
}
\description{
Create a \code{BOSTON} object to hold BOSTON scores.
}
