% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-class.R
\name{c.npsych_scores}
\alias{c.npsych_scores}
\title{Combine \code{npsych_scores} objects}
\usage{
\method{c}{npsych_scores}(x, ...)
}
\arguments{
\item{x}{An \code{npsych_scores} object.}

\item{...}{Additional \code{npsych_scores} objects to combine.}
}
\value{
An \code{npsych_scores} object containing the combined data from \code{x} and \code{...}.
The function will error if different subclasses of \code{npsych_scores} are provided.
}
\description{
Combine \code{npsych_scores} objects
}
\keyword{internal}
