% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_FAS.R
\name{calc_FAS}
\alias{calc_FAS}
\title{Calcaulate Functional Assessment Score summary}
\usage{
calc_FAS(
  BILLS,
  TAXES,
  SHOPPING,
  GAMES,
  STOVE,
  MEALPREP,
  EVENTS,
  PAYATTN,
  REMDATES,
  TRAVEL
)
}
\arguments{
\item{BILLS}{score for ability to write checks, pay bills, or balancing check book.}

\item{TAXES}{score for ability to assemble tax records, business affairs, or other papers.}

\item{SHOPPING}{score for shopping alone for clothes, household necessities, or groceries.}

\item{GAMES}{score for playing a game of skill such as bridge or chess, and working on a hobby}

\item{STOVE}{score for ability to heat water, make a cup of coffee, turn off the stove}

\item{MEALPREP}{score for ability to prepare a balanced meal}

\item{EVENTS}{score for keeping track of current events}

\item{PAYATTN}{score for ability to pay attention to and understand a TV program, a book, or a magazine}

\item{REMDATES}{score for remembering appointments, family occasions, holidays, and medications.}

\item{TRAVEL}{score for traveling out of the neighborhood, driving, or arranging to take public transportation}
}
\description{
Calculates the Functional Assessment Score summary (FAS) from the ten scores of
\itemize{
\item bills
\item taxes
\item shopping
\item games
\item stove
\item meal prep
\item events
\item paying attention
\item remembering appointments
\item travel
}

Each score is on the scale 0 (Normal), 1 (Has difficulty, but does by self),
2 (Requires assistance), 3 (dependent). Further, error codes of 8 (Not
applicable, e.g. never did) and 9 (Unknown) are also allowed, while -4
indicates data not collected. The FAS is calculated as the sum of the ten
scores where any 8s or 9s are counted as 0.
}
