% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MOCACLOCK.R
\name{calc_MOCACLOCK}
\alias{calc_MOCACLOCK}
\title{Calculate MoCA clock drawing test total}
\usage{
calc_MOCACLOCK(MOCACLOC, MOCACLON, MOCACLOH)
}
\arguments{
\item{MOCACLOC}{values for clock contour}

\item{MOCACLON}{values for clock numbers}

\item{MOCACLOH}{values for clock hands}
}
\description{
Sum of MOCACLOC, MOCACLON, and MOCACLOH, but applies removes any values not 0 or 1.
}
