% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REYAREC.R
\name{calc_REYAREC}
\alias{calc_REYAREC}
\title{Calculate the Rey AVLT Accuracy}
\usage{
calc_REYAREC(reytcor, reyfpos)
}
\arguments{
\item{reytcor}{true positives}

\item{reyfpos}{false positives}
}
\description{
Calculates accuracy from true positives and false positives, but applies \code{valid_values_only()}
first. As percentage, i.e. a value of 55 means 55\%.
}
