% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REYTOTAL.R
\name{calc_REYTOTAL}
\alias{calc_REYTOTAL}
\title{Calculate Rey Total}
\usage{
calc_REYTOTAL(rey1rec, rey2rec, rey3rec, rey4rec, rey5rec)
}
\arguments{
\item{rey1rec}{A \code{REY1REC} object, or a numeric vector.}

\item{rey2rec}{A \code{REY2REC} object, or a numeric vector.}

\item{rey3rec}{A \code{REY3REC} object, or a numeric vector.}

\item{rey4rec}{A \code{REY4REC} object, or a numeric vector.}

\item{rey5rec}{A \code{REY5REC} object, or a numeric vector.}
}
\value{
A \code{REYTOTAL} object representing the sum of the provided \code{rey} records
after error codes have been removed.
}
\description{
A short description...
}
