% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-CDRGLOB.R
\name{CDRGLOB}
\alias{CDRGLOB}
\title{CDRGLOB Test Scores}
\usage{
CDRGLOB(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CDRGLOB}.
}
\description{
Create a \code{CDRGLOB} object to hold CDRGLOB scores.
}
