% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-CDRSUM.R
\name{CDRSUM}
\alias{CDRSUM}
\title{CDRSUM Test Scores}
\usage{
CDRSUM(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CDRSUM}.
}
\description{
Create a \code{CDRSUM} object to hold CDRSUM scores.
}
