% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-CRAFTDRE.R
\name{CRAFTDRE}
\alias{CRAFTDRE}
\title{CRAFTDRE Test Scores}
\usage{
CRAFTDRE(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CRAFTDRE}.
}
\description{
Create a \code{CRAFTDRE} object to hold CRAFTDRE scores.
}
