% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-CRAFTDVR.R
\name{CRAFTDVR}
\alias{CRAFTDVR}
\title{CRAFTDVR Test Scores}
\usage{
CRAFTDVR(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CRAFTDVR}.
}
\description{
Create a \code{CRAFTDVR} object to hold CRAFTDVR scores.
}
