% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-CRAFTURS.R
\name{CRAFTURS}
\alias{CRAFTURS}
\title{CRAFTURS Test Scores}
\usage{
CRAFTURS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CRAFTURS}.
}
\description{
Create a \code{CRAFTURS} object to hold CRAFTURS scores.
}
