% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-CRAFTVRS.R
\name{CRAFTVRS}
\alias{CRAFTVRS}
\title{CRAFTVRS Test Scores}
\usage{
CRAFTVRS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{CRAFTVRS}.
}
\description{
Create a \code{CRAFTVRS} object to hold CRAFTVRS scores.
}
