% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGBACCT.R
\name{DIGBACCT}
\alias{DIGBACCT}
\title{DIGBACCT Test Scores}
\usage{
DIGBACCT(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGBACCT}.
}
\description{
Create a \code{DIGBACCT} object to hold DIGBACCT scores.
}
