% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGBACLS.R
\name{DIGBACLS}
\alias{DIGBACLS}
\title{DIGBACLS Test Scores}
\usage{
DIGBACLS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGBACLS}.
}
\description{
Create a \code{DIGBACLS} object to hold DIGBACLS scores.
}
