% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGFORCT.R
\name{DIGFORCT}
\alias{DIGFORCT}
\title{DIGFORCT Test Scores}
\usage{
DIGFORCT(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGFORCT}.
}
\description{
Create a \code{DIGFORCT} object to hold DIGFORCT scores.
}
