% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGFORSL.R
\name{DIGFORSL}
\alias{DIGFORSL}
\title{DIGFORSL Test Scores}
\usage{
DIGFORSL(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGFORSL}.
}
\description{
Create a \code{DIGFORSL} object to hold DIGFORSL scores.
}
