% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGIB.R
\name{DIGIB}
\alias{DIGIB}
\title{DIGIB Test Scores}
\usage{
DIGIB(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGIB}.
}
\description{
Create a \code{DIGIB} object to hold DIGIB scores.
}
