% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGIBLEN.R
\name{DIGIBLEN}
\alias{DIGIBLEN}
\title{DIGIBLEN Test Scores}
\usage{
DIGIBLEN(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGIBLEN}.
}
\description{
Create a \code{DIGIBLEN} object to hold DIGIBLEN scores.
}
