% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGIF.R
\name{DIGIF}
\alias{DIGIF}
\title{DIGIF Test Scores}
\usage{
DIGIF(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGIF}.
}
\description{
Create a \code{DIGIF} object to hold DIGIF scores.
}
