% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-DIGIFLEN.R
\name{DIGIFLEN}
\alias{DIGIFLEN}
\title{DIGIFLEN Test Scores}
\usage{
DIGIFLEN(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{DIGIFLEN}.
}
\description{
Create a \code{DIGIFLEN} object to hold DIGIFLEN scores.
}
