% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_std_methods.R
\name{.find_std_generics}
\alias{.find_std_generics}
\title{Look for generics matching std_using_* in all loaded namespaces}
\usage{
.find_std_generics()
}
\description{
Look for generics matching std_using_* in all loaded namespaces
}
\keyword{internal}
