% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-ANIMALS.R
\name{.setup_ANIMALS_versions}
\alias{.setup_ANIMALS_versions}
\title{Setup ANIMALS method versions}
\usage{
.setup_ANIMALS_versions()
}
\value{
\code{NULL}, invisibly. This function is called for its side effects of
registering test versions and setting defaults.
}
\description{
Registers versions for the ANIMALS test class,
and sets a default method. Meant to be called in .onLoad
}
\keyword{internal}
