% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-MOCATOTS.R,
%   R/test_scores-MOCATOTS.R
\name{.setup_MOCATOTS_versions}
\alias{.setup_MOCATOTS_versions}
\title{Setup MOCATOTS method versions}
\usage{
.setup_MOCATOTS_versions()

.setup_MOCATOTS_versions()
}
\value{
\code{NULL}, invisibly. This function is called for its side effects of
registering test versions and setting defaults.

\code{NULL}, invisibly. This function is called for its side effects of
registering test versions and setting defaults.
}
\description{
Registers versions for the MOCATOTS test class,
and sets a default method. Meant to be called in .onLoad

Registers versions for the MOCATOTS test class,
and sets a default method. Meant to be called in .onLoad
}
\keyword{internal}
