% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY6REC.R
\name{.setup_REY6REC_versions}
\alias{.setup_REY6REC_versions}
\title{Setup REY6REC method versions}
\usage{
.setup_REY6REC_versions()
}
\value{
\code{NULL}, invisibly. This function is called for its side effects of
registering test versions and setting defaults.
}
\description{
Registers versions for the \%s test class,
and sets a default method. Meant to be called in .onLoad
}
\keyword{internal}
