% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILA.R
\name{.setup_TRAILA_versions}
\alias{.setup_TRAILA_versions}
\title{Setup TRAILA method versions}
\usage{
.setup_TRAILA_versions()
}
\value{
\code{NULL}, invisibly. This function is called for its side effects of
registering test versions and setting defaults.
}
\description{
Registers versions for the \%s test class,
and sets a default method. Meant to be called in .onLoad
}
\keyword{internal}
