% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\name{get_default_method}
\alias{get_default_method}
\title{Get the default version for a method/test_class combination}
\usage{
get_default_method(test_class)
}
\arguments{
\item{test_class}{Character string or \code{test_scores} object with subclass.}
}
\value{
Character string of default version, or NULL if none set
}
\description{
Get the default version for a method/test_class combination
}
