% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npsych_scores.R
\name{get_npsych_scores}
\alias{get_npsych_scores}
\title{Get an npsych_scores constructor by name}
\usage{
get_npsych_scores(name)
}
\arguments{
\item{name}{A single string, e.g. "REYTOTAL".}
}
\value{
The S7 class object (which is callable as a constructor).
}
\description{
Get an npsych_scores constructor by name
}
