% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_defaults.R
\name{get_std_defaults}
\alias{get_std_defaults}
\title{Get standardization defaults \code{npsych_scores}}
\usage{
get_std_defaults(scores)
}
\arguments{
\item{scores}{S7 object of class \code{npsych_scores} with subclass.}
}
\value{
Character string with default method and version (when applicable)
}
\description{
Get standardization defaults \code{npsych_scores}
}
