% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std_methods.R
\name{get_std_methods}
\alias{get_std_methods}
\title{Get standardization methods for test scores}
\usage{
get_std_methods(test_class)
}
\arguments{
\item{test_class}{A \code{test_scores} object, such as \code{MOCATOTS()}. When the
argument is supplied, only methods with registered versions for that
specific test class are returned. When called via the \code{.test_scores}
method with a missing argument, all registered methods are returned.}
}
\value{
A character vector of standardization method names.
}
\description{
Get methods that are available for a test class, or all
available methods.
}
