% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-registry-functions.R
\name{get_test_score_classes}
\alias{get_test_score_classes}
\title{Get test score classes}
\usage{
get_test_score_classes()
}
\value{
A character vector of the names of test score classes.
}
\description{
Get all test score classes available.
}
\keyword{internal}
