% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_npsych_scores.R
\name{list_npsych_scores}
\alias{list_npsych_scores}
\title{List \code{npsych_scores} subclasses}
\usage{
list_npsych_scores()
}
\value{
A character vector of valid test subclass names.
}
\description{
This function is a helper that lists all available subclasses, i.e. all functions created using \code{new_class <- new_npsych_scores(...)}.
}
