% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_std_methods.R
\name{list_std_methods}
\alias{list_std_methods}
\title{List available standardization methods for \code{npsych_scores}}
\usage{
list_std_methods(scores)
}
\arguments{
\item{scores}{A \code{npsych_scores} object, such as \code{MOCATOTS()}. When the
argument is supplied, only methods with registered versions for that
specific test class are returned. When called via the \code{.npsych_scores}
method with a missing argument, all registered methods are returned.}
}
\value{
A character vector of standardization method names.
}
\description{
Get methods that are available for a test class, or all
available methods.
}
