% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-LOGIMEM.R
\name{LOGIMEM}
\alias{LOGIMEM}
\title{LOGIMEM Test Scores}
\usage{
LOGIMEM(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{LOGIMEM}.
}
\description{
Create a \code{LOGIMEM} object to hold LOGIMEM scores.
}
