% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-MEMUNITS.R
\name{MEMUNITS}
\alias{MEMUNITS}
\title{MEMUNITS Test Scores}
\usage{
MEMUNITS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{MEMUNITS}.
}
\description{
Create a \code{MEMUNITS} object to hold MEMUNITS scores.
}
