% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-MINTTOTS.R
\name{MINTTOTS}
\alias{MINTTOTS}
\title{MINTTOTS Test Scores}
\usage{
MINTTOTS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{MINTTOTS}.
}
\description{
Create a \code{MINTTOTS} object to hold MINTTOTS scores.
}
