% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-MOCACLOCK.R
\name{MOCACLOCK}
\alias{MOCACLOCK}
\title{MOCACLOCK Test Scores}
\usage{
MOCACLOCK(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{MOCACLOCK}.
}
\description{
Create a \code{MOCACLOCK} object to hold MOCACLOCK scores.
}
