% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-MOCBTOTS.R
\name{MOCBTOTS}
\alias{MOCBTOTS}
\title{MOCBTOTS Test Scores}
\usage{
MOCBTOTS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{MOCBTOTS}.
}
\description{
Create a \code{MOCBTOTS} object to hold MoCA Blind scores.
}
