% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-NACCGDS.R
\name{NACCGDS}
\alias{NACCGDS}
\title{NACCGDS Test Scores}
\usage{
NACCGDS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{NACCGDS}.
}
\description{
Create a \code{NACCGDS} object to hold Total GDS Scores.
}
