% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-NACCMMSE.R
\name{NACCMMSE}
\alias{NACCMMSE}
\title{NACCMMSE Test Scores}
\usage{
NACCMMSE(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{NACCMMSE}.
}
\description{
Create a \code{NACCMMSE} object to hold NACCMMSE scores.
}
