% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores_00.R
\name{new_test_scores}
\alias{new_test_scores}
\title{Create a \code{test_scores} class}
\usage{
new_test_scores(scores, label, range, codes, class)
}
\arguments{
\item{scores}{A numeric vector. Will error if not numeric.}

\item{label}{A single string.}

\item{range}{A numeric vector, typically of length two.}

\item{codes}{A vector or list of codes.}

\item{class}{Optional.}
}
\value{
A numeric vector with additional attributes \code{label}, \code{range}, \code{codes},
and classes \code{test_scores} and \code{class}. Raises an error if \code{scores} is
not a numeric vector.
}
\description{
Create a \code{test_scores} class. Used internally to create test specific classes.
}
\keyword{internal}
