% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrs-package.R
\docType{package}
\name{ntrs-package}
\alias{ntrs}
\alias{ntrs-package}
\title{ntrs: Neuropsychological Test Result Standardization}
\description{
S7 based framework for working with and standardizing neuropsychological test results. Built in tools for working with the NACC Uniform Data Set (UDS) and the National Alzheimer's Coordinating Center (NACC) neuropsychological battery, but with extensibility at the center of the design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmtrane.github.io/ntrs/}
  \item \url{https://github.com/rmtrane/ntrs}
  \item Report bugs at \url{https://github.com/rmtrane/ntrs/issues}
}

}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rmtrane@gmail.com} (\href{https://orcid.org/0000-0001-7824-1785}{ORCID})

}
\keyword{internal}
