% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-OTRAILA.R
\name{OTRAILA}
\alias{OTRAILA}
\title{OTRAILA Test Scores}
\usage{
OTRAILA(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{OTRAILA}.
}
\description{
Create a \code{OTRAILA} object to hold OTRAILA scores.
}
