% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-OTRAILB.R
\name{OTRAILB}
\alias{OTRAILB}
\title{OTRAILB Test Scores}
\usage{
OTRAILB(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{OTRAILB}.
}
\description{
Create a \code{OTRAILB} object to hold OTRAILB scores.
}
