% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-OTRLARR.R
\name{OTRLARR}
\alias{OTRLARR}
\title{OTRLARR Test Scores}
\usage{
OTRLARR(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{OTRLARR}.
}
\description{
Create a \code{OTRLARR} object to hold OTRLARR scores.
}
