% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-OTRLBRR.R
\name{OTRLBRR}
\alias{OTRLBRR}
\title{OTRLBRR Test Scores}
\usage{
OTRLBRR(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{OTRLBRR}.
}
\description{
Create a \code{OTRLBRR} object to hold OTRLBRR scores.
}
