% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\docType{data}
\name{rdd}
\alias{rdd}
\title{Researchers Data Dictionary in List Form}
\format{
\subsection{\code{rdd}}{

\describe{A list of entries for NACC variables with short descriptions,
allowable values, and codes. This is created from
from the .csv version of \href{https://files.alz.washington.edu/documentation/uds3-rdd.pdf}{this PDF},
which is made available to researcher upon acceptance of the data request,
with the addition of some entries calculated based on the NACC variables.
Each entry of the list has three entries:
\item{range}{For numeric variables, the smallest and largest values allowable (that are not error codes). For non-numeric variables, NULL}
\item{codes}{Named vector where entries are allowed codes and names are
labels. For example, for the total score for copy of Benson figure, the
value of \code{95} is in the vector with the name "Physical problem".}
\item{short_descriptor}{The short descriptor as given in the PDF linked above.}
}
}
}
\usage{
rdd
}
\description{
Researchers Data Dictionary in List Form
}
\keyword{datasets}
