% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores_00.R
\name{register_test_score_class}
\alias{register_test_score_class}
\title{Register a test score class}
\usage{
register_test_score_class(class_name)
}
\value{
\code{class_name}, invisibly. An error is thrown if \code{class_name} is not a
single character string.
}
\description{
A short description...
#' @param class_name A single character string.
}
