% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores.R
\name{remove_errorcodes}
\alias{remove_errorcodes}
\title{Remove error codes}
\usage{
remove_errorcodes(x)
}
\arguments{
\item{x}{An object of class \code{test_scores}}
}
\value{
An object with error codes removed.
}
\description{
Removes error codes from the \code{test_scores} object using the \code{codes} attribute.
}
