% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_version_registry.R
\name{remove_std_version}
\alias{remove_std_version}
\title{Remove a registered version}
\usage{
remove_std_version(test_class, method, version)
}
\arguments{
\item{test_class}{Character string identifying the test class}

\item{method}{Character string: "norms" or "regression"}

\item{version}{Character string identifying the version to remove}
}
\description{
Remove a registered version
}
