% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores_helpers.R
\name{replace_codes}
\alias{replace_codes}
\title{Replace codes with their labels}
\usage{
replace_codes(x)
}
\arguments{
\item{x}{An object of class \code{npsych_scores}.}
}
\value{
A numeric vector with code values replaced by their named labels.
}
\description{
Replaces numeric codes in a \code{npsych_scores} object with their corresponding
named labels from the \code{codes} attribute.
}
\examples{
moca <- MOCATOTS(c(25, 88, -4, 28))
replace_codes(moca)

}
