% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY1REC.R
\name{REY1REC}
\alias{REY1REC}
\title{REY1REC Test Scores}
\usage{
REY1REC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY1REC}.
}
\description{
Create a \code{REY1REC} object to hold REY1REC scores.
}
