% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY2REC.R
\name{REY2REC}
\alias{REY2REC}
\title{REY2REC Test Scores}
\usage{
REY2REC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY2REC}.
}
\description{
Create a \code{REY2REC} object to hold REY2REC scores.
}
