% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY3REC.R
\name{REY3REC}
\alias{REY3REC}
\title{REY3REC Test Scores}
\usage{
REY3REC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY3REC}.
}
\description{
Create a \code{REY3REC} object to hold REY3REC scores.
}
