% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY4REC.R
\name{REY4REC}
\alias{REY4REC}
\title{REY4REC Test Scores}
\usage{
REY4REC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY4REC}.
}
\description{
Create a \code{REY4REC} object to hold REY4REC scores.
}
