% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY5REC.R
\name{REY5REC}
\alias{REY5REC}
\title{REY5REC Test Scores}
\usage{
REY5REC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY5REC}.
}
\description{
Create a \code{REY5REC} object to hold REY5REC scores.
}
