% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REY6REC.R
\name{REY6REC}
\alias{REY6REC}
\title{REY6REC Test Scores}
\usage{
REY6REC(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REY6REC}.
}
\description{
Create a \code{REY6REC} object to hold REY6REC scores.
}
