% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REYAREC.R
\name{REYAREC}
\alias{REYAREC}
\title{REYAREC Test Scores}
\usage{
REYAREC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REYAREC}.
}
\description{
Create a \code{REYAREC} object to hold REYAREC scores.
}
