% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REYDLIST.R
\name{REYDLIST}
\alias{REYDLIST}
\title{REYDLIST Test Scores}
\usage{
REYDLIST(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REYDLIST}.
}
\description{
Create a \code{REYDLIST} object to hold REYDLIST scores.
}
