% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REYTCOR.R
\name{REYTCOR}
\alias{REYTCOR}
\title{REYTCOR Test Scores}
\usage{
REYTCOR(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REYTCOR}.
}
\description{
Create a \code{REYTCOR} object to hold REYTCOR scores.
}
