% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REYTOTAL.R
\name{REYTOTAL}
\alias{REYTOTAL}
\title{REYTOTAL Test Scores}
\usage{
REYTOTAL(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REYTOTAL}.
}
\description{
Create a \code{REYTOTAL} object to hold REYTOTAL scores.
}
