% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\name{set_default_method}
\alias{set_default_method}
\title{Set the default version for a method/test_class combination}
\usage{
set_default_method(test_class, method, version, overwrite = FALSE)
}
\arguments{
\item{test_class}{Character string identifying the test class}

\item{method}{Character string: "norms" or "regression"}

\item{version}{Character string identifying the version to set as default}

\item{overwrite}{Logical (default: \code{FALSE}); should existing default, if it exists, be overwritten?}
}
\description{
Set the default version for a method/test_class combination
}
