% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_defaults.R
\name{set_std_defaults}
\alias{set_std_defaults}
\title{Set the default standardization method (and version) an \code{npsych_scores} subclass}
\usage{
set_std_defaults(scores, method, version, overwrite = FALSE)
}
\arguments{
\item{scores}{S7 object of class \code{npsych_scores} with subclass.}

\item{method}{Character string, for example "norms" or "regression"}

\item{version}{Character string identifying the version to set as default}

\item{overwrite}{Logical (default: \code{FALSE}); should existing default, if it exists, be overwritten?}
}
\description{
Set the default standardization method (and version) an \code{npsych_scores} subclass
}
