% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_norms.R
\name{std_using_norms-npsych_scores}
\alias{std_using_norms-npsych_scores}
\title{Standardize test scores using norms}
\arguments{
\item{scores}{A numeric vector of raw test scores.}

\item{...}{Additional numeric vectors representing covariates, for example \code{age}, \code{educ}, \code{sex}.
These must be of length one or the same length as \code{npsych_scores}, and named as columns in the
\code{lookup_table} for the specified \code{version}; use \code{get_version_data({npsych_scores}, "norms", {version})$lookup_table}
to inspect the table.}

\item{version}{A single string specifying the version of the norms to use.}
}
\value{
A numeric vector of standardized scores. The function will error if \code{version} is not registered, if provided covariates are not numeric, if required covariates are missing based on the \code{lookup_table} for the specified \code{version}, or if covariate lengths are mismatched.
}
\description{
A short description...
}
