% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_norms.R
\name{std_using_norms}
\alias{std_using_norms}
\title{Standardize \code{test_scores} using norms}
\usage{
std_using_norms(test_scores, age, sex, educ, version, ...)
}
\arguments{
\item{age}{A numeric value.}

\item{sex}{A single string or numeric value.}

\item{educ}{A numeric value.}

\item{version}{A single string or numeric value.}

\item{...}{Arguments passed to methods.}

\item{x}{A numeric vector.}
}
\value{
The standardized value of \code{x}, whose exact type depends on the specific S3
method called.
}
\description{
A short description...
}
