% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_regression.R
\name{std_using_regression-npsych_scores}
\alias{std_using_regression-npsych_scores}
\title{Standardize test scores using regression}
\arguments{
\item{scores}{A vector coercible to numeric, representing raw test scores.}

\item{...}{Additional numeric covariates to be included in the regression model. Each covariate must be a numeric vector of length 1 or the same length as \code{npsych_scores},
and must be named after a coefficient present in the \code{version} data; use \code{get_version_data({npsych_scores}, "regression", {version})$coefs} to inspect the coefficients.}

\item{version}{A single string specifying the version of the regression model to use for standardization.}
}
\value{
A numeric vector of standardized test scores. The function will error if any supplied covariates are not numeric, if required covariates for the specified \code{version} are missing, or if covariate lengths are mismatched.
}
\description{
A short description...
}
