% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores-MOCATOTS.R
\name{std_using_regression.MOCATOTS}
\alias{std_using_regression.MOCATOTS}
\title{Standardize MOCATOTS scores using regression}
\usage{
\method{std_using_regression}{MOCATOTS}(test_scores, age, sex, educ, race, version)
}
\arguments{
\item{test_scores}{A numeric vector of raw MOCATOTS scores of class \code{MOCATOTS}.}

\item{age}{A numeric vector of ages.}

\item{sex}{A character vector of sexes.}

\item{educ}{A numeric vector of education levels.}

\item{race}{A character vector of races.}

\item{version}{A single string indicating the version of the test.}
}
\value{
A numeric vector of standardized scores.
}
\description{
Standardize MOCATOTS scores using regression
}
