% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_regression.R
\name{std_using_regression}
\alias{std_using_regression}
\title{Standardize using regression}
\usage{
std_using_regression(test_scores, ...)
}
\arguments{
\item{test_scores}{A numeric vector of test scores.}

\item{...}{Arguments passed to the specific method.}
}
\value{
The result of the dispatched method for \code{std_using_regression}.
}
\description{
A short description...
}
