% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores_00.R
\name{test_scores}
\alias{test_scores}
\title{Test scores}
\usage{
test_scores(
  scores,
  label = rdd[[class]]$short_descriptor,
  range = rdd[[class]]$range,
  codes = rdd[[class]]$codes,
  class = character()
)
}
\arguments{
\item{scores}{A numeric vector.}

\item{label}{A single string. Optional. Defaults to \code{short_descriptor} in \code{rdd}.}

\item{range}{A numeric vector of length 2. Optional. Defaults to \code{range} in \code{rdd}.}

\item{codes}{A character vector. Optional. Defaults to \code{codes} in \code{rdd}.}

\item{class}{A single string. Optional. Additional classes to add to the resulting class.}
}
\value{
A validated \code{test_scores} object. Will error if validation fails.
}
\description{
A short description...
}
