% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-class.R,
%   R/std_npsych_scores-class.R
\name{`[`}
\alias{`[`}
\title{Subset \code{npsych_scores} objects.}
\arguments{
\item{x}{An S7 object of class \code{npsych_scores}.}

\item{i}{A numeric index.}
}
\value{
A new \code{npsych_scores} object containing the elements of \code{x} specified by \code{i}.

A new \code{npsych_scores} object containing the elements of \code{x} specified by \code{i}.
}
\description{
Subset \code{npsych_scores} objects.

Subset \code{npsych_scores} objects.
}
\keyword{internal}
