% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILA.R
\name{TRAILA}
\alias{TRAILA}
\title{TRAILA Test Scores}
\usage{
TRAILA(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{TRAILA}.
}
\description{
Create a \code{TRAILA} object to hold TRAILA scores.
}
