% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILALI.R
\name{TRAILALI}
\alias{TRAILALI}
\title{TRAILALI Test Scores}
\usage{
TRAILALI(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{TRAILALI}.
}
\description{
Create a \code{TRAILALI} object to hold TRAILALI scores.
}
