% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILARR.R
\name{TRAILARR}
\alias{TRAILARR}
\title{TRAILARR Test Scores}
\usage{
TRAILARR(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{TRAILARR}.
}
\description{
Create a \code{TRAILARR} object to hold TRAILARR scores.
}
