% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILB.R
\name{TRAILB}
\alias{TRAILB}
\title{TRAILB Test Scores}
\usage{
TRAILB(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{TRAILB}.
}
\description{
Create a \code{TRAILB} object to hold TRAILB scores.
}
