% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILBLI.R
\name{TRAILBLI}
\alias{TRAILBLI}
\title{TRAILBLI Test Scores}
\usage{
TRAILBLI(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{TRAILBLI}.
}
\description{
Create a \code{TRAILBLI} object to hold TRAILBLI scores.
}
