% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-TRAILBRR.R
\name{TRAILBRR}
\alias{TRAILBRR}
\title{TRAILBRR Test Scores}
\usage{
TRAILBRR(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{TRAILBRR}.
}
\description{
Create a \code{TRAILBRR} object to hold TRAILBRR scores.
}
