% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-UDSBENRS.R
\name{UDSBENRS}
\alias{UDSBENRS}
\title{UDSBENRS Test Scores}
\usage{
UDSBENRS(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{UDSBENRS}.
}
\description{
Create a \code{UDSBENRS} object to hold UDSBENRS scores.
}
