% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-UDSBENTC.R
\name{UDSBENTC}
\alias{UDSBENTC}
\title{UDSBENTC Test Scores}
\usage{
UDSBENTC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{UDSBENTC}.
}
\description{
Create a \code{UDSBENTC} object to hold UDSBENTC scores.
}
