% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-UDSBENTD.R
\name{UDSBENTD}
\alias{UDSBENTD}
\title{UDSBENTD Test Scores}
\usage{
UDSBENTD(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{UDSBENTD}.
}
\description{
Create a \code{UDSBENTD} object to hold UDSBENTD scores.
}
