% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-UDSVERFC.R
\name{UDSVERFC}
\alias{UDSVERFC}
\title{UDSVERFC Test Scores}
\usage{
UDSVERFC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{UDSVERFC}.
}
\description{
Create a \code{UDSVERFC} object to hold UDSVERFC scores.
}
