% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-UDSVERLC.R
\name{UDSVERLC}
\alias{UDSVERLC}
\title{UDSVERLC Test Scores}
\usage{
UDSVERLC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{UDSVERLC}.
}
\description{
Create a \code{UDSVERLC} object to hold UDSVERLC scores.
}
