% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-UDSVERTN.R
\name{UDSVERTN}
\alias{UDSVERTN}
\title{UDSVERTN Test Scores}
\usage{
UDSVERTN(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{UDSVERTN}.
}
\description{
Create a \code{UDSVERTN} object to hold UDSVERTN scores.
}
