% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores.R
\name{validate_test_scores}
\alias{validate_test_scores}
\title{Validate test scores}
\usage{
validate_test_scores(x)
}
\arguments{
\item{x}{An object representing test scores created using \code{test_scores()}. It must be coercible
to a numeric vector
and have the following attributes: \code{label} (a single string), \code{range} (a numeric vector
of length 2), and \code{codes} (a named numeric vector). Its class (excluding \code{"test_scores"})
must be a single character string.}
}
\value{
\code{x}, invisibly, if all validations pass. The function will raise an error if:
\itemize{
\item The class of \code{x} (excluding \code{"test_scores"}) is not a single character string.
\item The \code{label} attribute is not a single string.
\item The \code{range} attribute is not a numeric vector of length 2.
\item The \code{codes} attribute is not a named numeric vector.
\item Any score in \code{x} is not within the specified \code{range} and not among the \code{codes}.
}
}
\description{
Used to validate the creation of new \code{test_scores}.
}
\keyword{internal}
