% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-VEG.R
\name{VEG}
\alias{VEG}
\title{VEG Test Scores}
\usage{
VEG(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{VEG}.
}
\description{
Create a \code{VEG} object to hold VEG scores.
}
