% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-WAIS.R
\name{WAIS}
\alias{WAIS}
\title{WAIS Test Scores}
\usage{
WAIS(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{WAIS}.
}
\description{
Create a \code{WAIS} object to hold WAIS scores.
}
