% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-registry-functions.R
\name{.register_std_version}
\alias{.register_std_version}
\title{Register a version}
\usage{
.register_std_version(
  test_class,
  method,
  version,
  data,
  description = "",
  overwrite = FALSE
)
}
\arguments{
\item{test_class}{A single string.}

\item{method}{A single string. Must be one of the implemented methods. See \code{get_}}

\item{version}{A single string.}

\item{data}{Data.}

\item{description}{A single string. Optional.}

\item{overwrite}{A logical value. Optional.}
}
\value{
Returns \code{NULL}, invisibly. The function will raise an error if \code{method} is not
implemented for \code{test_class} or if \code{version} already exists and \code{overwrite} is \code{FALSE}.
A warning will be issued if \code{version} exists and \code{overwrite} is \code{TRUE}.
}
\description{
A short description...
}
\keyword{internal}
