% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_data.R
\name{get_version_data}
\alias{get_version_data}
\title{Get version data}
\usage{
get_version_data(test_class, method = "default", version = NULL)
}
\arguments{
\item{test_class}{A single string specifying the test class.}

\item{method}{A single string specifying the method.}

\item{version}{A single string. If \code{"default"} (default), the function will attempt to retrieve the default version for the given method and test class.}
}
\value{
The object representing the requested version data. The function will error if \code{version} is \code{"default"} and no default version is set,
or if the specified \code{version} (or the resolved default version) does not exist for the given \code{method} and \code{test_class}.
}
\description{
A short description...
}
