% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_version_registry.R
\name{get_versions}
\alias{get_versions}
\title{Get a specific standardization version or all available versions}
\usage{
get_versions(test_class, method)
}
\arguments{
\item{test_class}{Character string identifying the test class}

\item{method}{Character string giving method.}

\item{version}{Character string identifying the version. If NULL, uses default.}
}
\value{
List containing version information and data
}
\description{
Get a specific standardization version or all available versions
}
