% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores_helpers.R
\name{is_npsych_scores}
\alias{is_npsych_scores}
\title{Is \code{x} an npsych_scores object?}
\usage{
is_npsych_scores(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A single logical value: \code{TRUE} if \code{x} is an \code{npsych_scores} object, \code{FALSE} otherwise.
}
\description{
Tests whether \code{x} is an \code{npsych_scores} object.
}
\examples{
moca <- MOCATOTS(c(25, 28))
is_npsych_scores(moca) # TRUE

traila <- c(53, 75)
is_npsych_scores(traila) # FALSE


}
