% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-MOCATOTS.R
\name{MOCATOTS}
\alias{MOCATOTS}
\title{MOCATOTS Test Scores}
\usage{
MOCATOTS(x = integer())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{MOCATOTS}.
}
\description{
Create a \code{MOCATOTS} object to hold MoCA scores.
}
