% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_version_registry.R
\name{register_norms_version}
\alias{register_norms_version}
\title{Register a norms-based standardization version}
\usage{
register_norms_version(
  test_class,
  version,
  lookup_table,
  covariate_prep_funs,
  description = "",
  overwrite = FALSE
)
}
\arguments{
\item{test_class}{Character string identifying the test class (e.g., "MOCATOTS")}

\item{version}{Character string identifying this version (e.g., "nacc", "updated")}

\item{lookup_table}{Data frame containing norms with columns: age_group, mean, sd}

\item{description}{Optional description of this version}

\item{age_group_fun}{Function taking a numeric vector as argument and returns
exactly the factor from \code{lookup_table$age_group}, if this is present. Internal
check is performed by comparing \code{unique(age_group_fun(0:110))} to \code{unique(lookup_table$age_group)}
using \code{all.equal}.}

\item{educ_group_fun}{Function taking a numeric vector as argument and returns
exactly the factor from \code{lookup_table$educ_group}, if this is present. Internal
check is performed by comparing \code{unique(educ_group_fun(0:50))} to \code{unique(lookup_table$educ_group)}
using \code{all.equal}.}
}
\description{
Register a norms-based standardization version
}
