% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_version_registry.R
\name{register_regression_version}
\alias{register_regression_version}
\title{Register a regression-based standardization version}
\usage{
register_regression_version(
  test_class,
  version,
  coefs,
  covariate_prep_funs,
  description = ""
)
}
\arguments{
\item{test_class}{Character string identifying the test class}

\item{version}{Character string identifying this version}

\item{description}{Optional description of this version}
}
\description{
Register a regression-based standardization version
}
