% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores_helpers.R
\name{remove_error_codes}
\alias{remove_error_codes}
\title{Remove error codes}
\usage{
remove_error_codes(x)
}
\arguments{
\item{x}{An object of class \code{npsych_scores}}
}
\value{
An object with error codes removed.
}
\description{
Removes error codes from the \code{npsych_scores} object using the \code{codes} attribute.
}
\examples{
# 88 is an error code for MOCATOTS (outside range 0-30)
moca <- MOCATOTS(c(25, 88, 28))
remove_error_codes(moca) # 25 NA 28

}
