% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsych_scores-REYDREC.R
\name{REYDREC}
\alias{REYDREC}
\title{REYDREC Test Scores}
\usage{
REYDREC(x = numeric())
}
\arguments{
\item{x}{Numeric scores.}
}
\value{
An object of class \code{REYDREC}.
}
\description{
Create a \code{REYDREC} object to hold REYDREC scores.
}
