% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_norms.R
\name{std_using_norms}
\alias{std_using_norms}
\title{Standardize \code{npsych_scores} using norms}
\usage{
std_using_norms(scores, ...)
}
\arguments{
\item{scores}{A numeric vector.}

\item{...}{Arguments passed to methods.}
}
\value{
The standardized value of \code{x}, whose exact type depends on the specific S3
method called.
}
\description{
A short description...
}
