% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_using_regression.R
\name{std_using_regression}
\alias{std_using_regression}
\title{Standardize using regression}
\usage{
std_using_regression(test_scores, age, sex, educ, version, ...)
}
\arguments{
\item{test_scores}{A numeric vector of test scores.}

\item{age}{A numeric vector for age.}

\item{sex}{A character vector for sex.}

\item{educ}{A numeric vector for education level.}

\item{version}{A single string or number indicating the version.}

\item{...}{Arguments passed to the specific method.}

\item{race}{A character vector for race.}
}
\value{
The result of the dispatched method for \code{std_using_regression}.
}
\description{
A short description...
}
