% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores.R
\name{test_scores}
\alias{test_scores}
\title{Test scores}
\usage{
test_scores(
  scores = numeric(),
  label,
  range,
  codes = numeric(),
  class = character()
)
}
\arguments{
\item{scores}{A numeric vector.}

\item{label}{A single string.}

\item{range}{A numeric vector of length 2.}

\item{codes}{A character vector. Optional. Should be a named character vector of the form \verb{c("error_label" = \{numeric code\})}.}

\item{class}{A single string. Optional.}
}
\value{
A validated \code{test_scores} object. Will error if validation fails.
}
\description{
A short description...
}
